Hdfs: Rearrange module structure
