+++ hdfs/src/main/scala/akka/stream/alpakka/hdfs/impl/writer/SequenceWriter.scala
@@ -2,10 +2,10 @@
  * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
  */
 
-package akka.stream.alpakka.hdfs.writer
+package akka.stream.alpakka.hdfs.impl.writer
 
-import akka.stream.alpakka.hdfs.scaladsl.FilePathGenerator
-import akka.stream.alpakka.hdfs.writer.HdfsWriter.createTargetPath
+import akka.stream.alpakka.hdfs.FilePathGenerator
+import akka.stream.alpakka.hdfs.impl.writer.HdfsWriter.createTargetPath
 import org.apache.hadoop.fs.{FileSystem, Path}
 import org.apache.hadoop.io.SequenceFile.{CompressionType, Writer}
 import org.apache.hadoop.io.compress.CompressionCodec
