--- hdfs/src/main/scala/akka/stream/alpakka/hdfs/scaladsl/Strategy.scala
@@ -1,12 +1,0 @@
-/*
- * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
- */
-
-package akka.stream.alpakka.hdfs.scaladsl
-
-private[scaladsl] trait Strategy {
-  type S <: Strategy
-  def should(): Boolean
-  def reset(): S
-  def update(offset: Long): S
-}
